464 added, 112 modified, 15 deleted C:\save\modding\ervin_mod_1

+++ src/main/java/com/babcsany/minecraft/ervin_mod_1/ervin_mod_1/interfaces/util/Maps.java
@@ -7,7 +7,7 @@ import java.util.LinkedHashMap;
 
 public interface Maps {
 
-    static <K, V> Map<Item, Double> newLinkedHashMap() {
-        return (Map<Item, Double>) new LinkedHashMap<K, V>();
+    static <K, V> Map<Item, Integer> newLinkedHashMap() {
+        return (Map<Item, Integer>) new LinkedHashMap<K, V>();
     }
 }
